7. There are only 3 colors, 10 digits, and 7 notes; it's what we do with them that's important

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -9,7 +9,7 @@
 
 int print_last_digit(int n)
 {
-	int last_digit = (unsigned int) n % 10;
+	int last_digit = (int) (unsigned int) n % 10;
 
 	_putchar('0' + last_digit);
 	return (last_digit);
